COM-3739: move tabsNames
